대상 분할 및 추가

+++ src/main/java/project/file/Item.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Data
 public class Item {
 
-    private Long itemId;
+    private Integer itemId; // 1이면 경찰, 2면 수능감독관, 3이면 토익감독관, 4면 비회원
     private List<UploadFile> imageFiles;
 }
