+++ src/main/java/project/file/ItemForm.java
@@ -8,6 +8,6 @@ import java.util.List;
 @Data
 public class ItemForm {
 
-    private Long itemId;
+    private Integer itemId; // 1이면 경찰, 2면 수능감독관, 3이면 토익감독관, 4면 비회원
     private List<MultipartFile> imageFiles;
 }
